refactor: Delegate system and VSCode backup logic to dedicated modules.
